LayoutTests/imported/w3c/web-platform-tests/content-security-policy/prefetch-src/prefetch-allowed-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'prefetch-src'.


Harness Error (TIMEOUT), message = null

PASS Browser supports prefetch.
PASS Browser supports performance APIs.
TIMEOUT Prefetch succeeds when allowed by prefetch-src Test timed out
